fix(log): concurrence problem of `Filter`
